Added a group size input instead of always having pairs
Also renamed a lot of variables to be accurate with groups (instead of pairs).
The minimal group size is 2 and the max group size is total participants - 2 so at least 2 groups can be formed and no one is alone. I also added a prompt to delete the previous pairings for ease of testing since it'll fill up quickly.